Move and dedup utils to internal
